feat(users): :sparkles: Add roles table and field to Breeze system
Created controllers, models, views, and middlewares necessary to implement user role management in Laravel Breeze system. This allows for assigning different permissions and privileges to users based on their roles in the system.

Added a new "roles" table to the database and added a new "role_id" field to the "users" table.  In