feat(users): create user management page with table and pagination
This commit adds a new feature to the application by creating a user management page. The page includes a table that displays all the users currently in the system, and pagination functionality has been implemented using the Laravel-Vue-Pagination library. This will allow the user to easily navigate through the list of users, as well as quickly find the information they are looking for.