feat(config): Move env variables to config file
This commit moves the env variables used in the DatabaseSeeder file to the
config file, in line with best practices. This change increases the security
of the application and makes it easier to manage environment variables.

BREAKING CHANGE: The environment variables in the DatabaseSeeder file have
been moved to the config file. Any references to these variables in the
DatabaseSeeder file must be updated to use the new config values.