fix: ensure correct storage directory permissions and update seeder data
Previously, the ProductFactory was unable to save generated images to the correct directory due to incorrect directory permissions. This commit fixes the issue by ensuring the necessary write permissions are set on the storage directory.

In addition, the number of brands and categories created by the seeders has been updated to better reflect the needs of the application.

This commit also includes updates to the relevant tests that were affected by the changes to the seeders and storage directory permissions.