feat(products): implement fake disk for testing and update ProductFactory
The purpose of this commit is to implement a fake disk for testing purposes related to the products feature. The ProductFactory has also been modified to use the disk depending on whether tests are being run or not. Additionally, the use of Storage::fake() has been added for tests that involve saving images.