Title: Add page to edit individual product, create necessary Vue views and ProductController
The necessary Vue views were created to render the page for editing an individual product. In addition, a ProductController was created to handle the data retrieval and rendering of the view. This feature should make it easier for users to edit individual products without having to navigate through the entire product list.