feat: Remove ImageController and create symbolic link to /public
The ImageController has been removed from the codebase as it was no longer necessary. Instead, a symbolic link to the /public directory has been created using the `php artisan storage:link` command. This will allow access to the images stored in the /storage directory from the /public directory.

Changes Made:
- Removed ImageController
- Created symbolic link to /public using `php artisan storage:link`