feat: refactor ProductController and add ProductImageService
This commit includes a series of changes to the ProductController and introduces a new ProductImageService.

The ProductImageService is now responsible for handling the storage and deletion of product images, and the ProductController has been updated to use this new service in its store and update methods.

Additionally, the responsibility for updating product status has been moved to a new ProductStatusController. The ProductController's update method no longer handles status updates, and instead this is handled by the update method of the ProductStatusController.