feat: Improve product table view to eliminate redundant server requests
Previously, the product table view suffered from the same issue as the user table view, requiring additional server requests to update the displayed data. This commit addresses the problem by optimizing the product table view, removing the need for extra requests. As a result, redundant controllers are eliminated, and the rendering of the view is unified by passing the data to be displayed.