feat: Combine Web/HomeController and Api/HomeController
Previously, the data for rendering the view required an additional request after the page was already rendered. This commit unifies the Web/HomeController and Api/HomeController, allowing the data to be passed along with the view, eliminating the need for a second request. This change also involves updating the associated views and tests.